Clear and delete the chain that was created in the tests.
In the TestRules test method, there was a chain being created to insert
rules into. Legit. There was no cleaning of this chain after the rules tests
were completed. Not legit. This meant that every time you would run the tests,
it would leave behind a TEST-<rand int> chain. Now, as the last 2 parts of that
method, it will clear the chain and then delete it.