new iptables constructor
Use functional options to create a new iptables object, so we can
support new options to configure it.

It allows to configure the IP family and the timeout used in the
`-w` flag. Until now, if iptables version supports wait,
we used `-w` without any timeout, so it keeps waiting forever trying
to acquire the lock at a 1 second interval rate.
This can cause issues on busy environments by software that depends
on this library, because they can be waiting forever.

Signed-off-by: Antonio Ojea <aojea@redhat.com>